json webtoken initialized
