+++ server/routes/user.routes.js
@@ -4,13 +4,12 @@ import {
   getUserById,
   updateUser,
   deleteUser,
-} from '../controllers/user.controller.js';
-
-import {
   createUser,
   otpverify,
   loginUser,
-} from '../controllers/user.auth.js';
+} from '../controllers/user.controller.js';
+
+
 
 const router = express.Router();
 
@@ -18,9 +17,9 @@ const router = express.Router();
 
 router.get('/users', getAllUsers);
 router.get('/user/:id', getUserById);
-router.post('/add-user', createUser);
+router.post('/signup', createUser);
 router.post('/login',loginUser)
-router.post('/otp-verfy',otpverify)
+router.post('/verify-otp',otpverify)
 router.put('/user/:id', updateUser);
 router.delete('/user/:id', deleteUser);
 export default router;
